completed encryption, also user can have multiple encryption key to hide multiple nortes, means a set of notes (emotions) would be discplayed when used the correct key.
Key 1 is going to show xyz emotions
Key 2 is going to show abc emotions

and if we have replaced key, we can use the same key back if we havent removed the data from the app. so different keys for different emotions can be used if the user wants. Currently implemented in offline mode.